Organized files and added ML notes
